requirements.txt
# Core dependencies
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=10.0.0

# Web interface
streamlit>=1.28.0

# Configuration and utilities
PyYAML>=6.0
pathlib2>=2.3.7

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: GPU support (uncomment if using CUDA)
# torch>=2.0.0
# torchvision>=0.15.0

# Optional: Additional ML libraries
# scikit-learn>=1.3.0
# pandas>=2.0.0
